Port of CorsixTH to SDL 3.4+

This is a conservative port, with only minimal changes and an effort to preserve the SDL2 look and feel.

Some distros ship SDL 3.4 but do not yet ship SDL3_mixer - so a FETCH_SDL_MIXER option was added to make it easy to build there, alternately vcpkg works well even on older distros.

There is a warning message printed to the console at least on my system that appears to be a regression in SDL_Mixer: libsdl-org/SDL_mixer#875 but it does not cause any problems running the game.

On Windows with sf3 soundfonts (particularly larger sf3 soundfonts) there is a noticeable delay between music tracks. In the short term using an sf2 soundfont e.g. https://musical-artifacts.com/artifacts/1229 is the recommended workaround.

It is noteworthy that the wav parser in SDL3_mixer is more strict about malformed files than SDL2_mixer was. Certain unofficial versions of Theme Hospital are known not to work, but retail versions appear fine as far as we've seen.

A few efforts are outside of the scope of this PR but should be followed up on afterwards:

    HiDPI support - A key feature of SDL3 was improved HiDPI support and we should take advantage.
    Fluidsynth - We could make a better user experience by taking more control of our fluidsynth playback rather than using SDL3_mixer.
    Further refactoring - I saw lots of opportunities to refactor in the process of making this patch series and had to resist to keep it small.
        sdl_audio.cpp should be merged with th_sound.cpp
        SDL_SetLogicalRenderingPresentation is used unnecessarily in the zoom code
        some form of dependency injection would be better than globals (fps / sdl_mixer) or throwing lots of singleton classes in lua registers.

+++ CorsixTH/Src/th_gfx.h
@@ -323,7 +323,7 @@ class animation_manager {
       @param scale_factor
   */
   void draw_frame(render_target* pCanvas, size_t iFrame,
-                  const ::layers& oLayers, int iX, int iY, uint32_t iFlags,
+                  const ::layers& oLayers, float iX, float iY, uint32_t iFlags,
                   animation_effect patient_effect = animation_effect::none,
                   size_t patient_effect_offset = 0, int scale_factor = 1) const;
